pyproject.toml
[project]
name = "learning-design-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio-tools>=1.76.0",
]
